update pay to lvye.

+++ pub_site/src/manager.py
@@ -30,7 +30,8 @@ def migrate():
 
 
 @manager.option('-r', '--recreate', action="store_true", dest="recreate", required=False, default=False)
-def init_db(recreate):
+@manager.option('-d', '--drop_all', action="store_true", dest="drop_all", required=False, default=False)
+def init_db(recreate, drop_all):
     from fabric.api import local
     from pub_site import db
 
@@ -42,7 +43,8 @@ def init_db(recreate):
 
     if recreate:
         recreate_db()
-    db.drop_all()
+    if drop_all:
+        db.drop_all()
     db.create_all()
